requirements.txt
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
pandas==2.1.4
openpyxl==3.1.2
click==8.1.7
pydantic==2.11.7
pydantic-settings==2.10.1
loguru==0.7.2 